+++ lib/Components/FilterBar/FilterBar.jsx
@@ -1,4 +1,5 @@
 import WrittenOnlyToggle from './WrittenOnlyToggle';
+import CollapseAllButton from './CollapseAllButton';
 
 import './FilterBar.scss';
 
@@ -8,6 +9,7 @@ export default function FilterBar({ variables }) {
       <div className="bio-vo-filter-bar-actions">
         <WrittenOnlyToggle />
       </div>
+      <CollapseAllButton />
     </div>
   );
 }