+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to the [@bpmn-io/variable-outline](https://github.com/bpmn-i
 * `FEAT`: add copy button for variable name ([#23](https://github.com/bpmn-io/variable-outline/pull/23))
 * `FEAT`: add written by selection filter ([#21](https://github.com/bpmn-io/variable-outline/pull/21))
 * `FEAT`: show variable value with codemirror ([#24](https://github.com/bpmn-io/variable-outline/pull/24))
+* `FEAT`: add collapse/expand scope groups functionality ([#22](https://github.com/bpmn-io/variable-outline/pull/22))
 
 ## 2.0.1